Split service.events into forge-specific sub-modules
Fixes #672
