+++ packit_service/service/events/__init__.py
@@ -1,0 +1,59 @@
+# Copyright Contributors to the Packit project.
+# SPDX-License-Identifier: MIT
+
+from packit_service.service.events.copr import (
+    AbstractCoprBuildEvent,
+    CoprBuildStartEvent,
+    CoprBuildEndEvent,
+)
+from packit_service.service.events.distgit import DistGitCommitEvent
+from packit_service.service.events.event import EventData, Event
+from packit_service.service.events.github import (
+    InstallationEvent,
+    IssueCommentEvent,
+    PullRequestCommentGithubEvent,
+    PullRequestGithubEvent,
+    PushGitHubEvent,
+    ReleaseEvent,
+    AbstractGithubEvent,
+)
+from packit_service.service.events.gitlab import (
+    IssueCommentGitlabEvent,
+    MergeRequestCommentGitlabEvent,
+    MergeRequestGitlabEvent,
+    PushGitlabEvent,
+)
+from packit_service.service.events.koji import KojiBuildEvent
+from packit_service.service.events.pagure import (
+    PullRequestPagureEvent,
+    PullRequestCommentPagureEvent,
+    PushPagureEvent,
+    AbstractPagureEvent,
+)
+from packit_service.service.events.testing_farm import TestingFarmResultsEvent
+
+__all__ = [
+    Event.__name__,
+    EventData.__name__,
+    AbstractCoprBuildEvent.__name__,
+    CoprBuildStartEvent.__name__,
+    CoprBuildEndEvent.__name__,
+    DistGitCommitEvent.__name__,
+    AbstractGithubEvent.__name__,
+    PushGitHubEvent.__name__,
+    PullRequestGithubEvent.__name__,
+    PullRequestCommentGithubEvent.__name__,
+    IssueCommentEvent.__name__,
+    InstallationEvent.__name__,
+    ReleaseEvent.__name__,
+    PushGitlabEvent.__name__,
+    MergeRequestGitlabEvent.__name__,
+    MergeRequestCommentGitlabEvent.__name__,
+    IssueCommentGitlabEvent.__name__,
+    KojiBuildEvent.__name__,
+    AbstractPagureEvent.__name__,
+    PushPagureEvent.__name__,
+    PullRequestCommentPagureEvent.__name__,
+    PullRequestPagureEvent.__name__,
+    TestingFarmResultsEvent.__name__,
+]
