+++ packit_service/worker/result.py
@@ -4,6 +4,7 @@
 from typing import Any, Dict
 
 from packit.config import JobConfig
+
 from packit_service.service.events import Event
 from packit_service.utils import dump_job_config, dump_package_config
