+++ tests/unit/test_events.py
@@ -22,32 +22,34 @@ from packit_service.models import (
     KojiBuildModel,
     TestingFarmResult,
     TFTTestRunModel,
+    AllowlistStatus,
 )
 from packit_service.service.db_triggers import AddPullRequestDbTrigger
 from packit_service.service.events import (
-    AllowlistStatus,
-    InstallationEvent,
-    ReleaseEvent,
-    PullRequestGithubEvent,
-    PullRequestAction,
-    TestingFarmResultsEvent,
-    PullRequestCommentGithubEvent,
-    PullRequestCommentAction,
-    IssueCommentEvent,
-    IssueCommentAction,
-    AbstractCoprBuildEvent,
-    FedmsgTopic,
-    DistGitCommitEvent,
-    PushGitHubEvent,
-    PullRequestPagureEvent,
-    PullRequestCommentPagureEvent,
-    MergeRequestGitlabEvent,
-    GitlabEventAction,
+    EventData,
     KojiBuildEvent,
+    TestingFarmResultsEvent,
+    DistGitCommitEvent,
+    AbstractCoprBuildEvent,
+    PullRequestPagureEvent,
     IssueCommentGitlabEvent,
     MergeRequestCommentGitlabEvent,
+    MergeRequestGitlabEvent,
     PushGitlabEvent,
-    EventData,
+    InstallationEvent,
+    IssueCommentEvent,
+    PullRequestCommentGithubEvent,
+    PullRequestGithubEvent,
+    PushGitHubEvent,
+    ReleaseEvent,
+    PullRequestCommentPagureEvent,
+)
+from packit_service.service.events.enums import (
+    PullRequestAction,
+    PullRequestCommentAction,
+    IssueCommentAction,
+    FedmsgTopic,
+    GitlabEventAction,
 )
 from packit_service.worker.parser import Parser, CentosEventParser
 from packit_service.worker.testing_farm import TestingFarmJobHelper
